requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.3
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
email-validator
python-dotenv==1.0.1
